refactor: move pkg/common to /common and pkg/core/ptr to core/ptr
