+++ common/entity.go
