+++ common/failure_mode.go
