+++ core/cipher/cipher.go
@@ -2,7 +2,7 @@
 package cipher
 
 import (
-	"github.com/aawadallak/go-core-kit/pkg/common"
+	"github.com/aawadallak/go-core-kit/common"
 )
 
 type TypeErrInvalidHash struct {
