+++ core/idem/manager_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/aawadallak/go-core-kit/core/idem"
-	"github.com/aawadallak/go-core-kit/pkg/common"
+	"github.com/aawadallak/go-core-kit/common"
 	"github.com/aawadallak/go-core-kit/plugin/idem/inmem"
 )
