+++ core/job/job.go
@@ -6,7 +6,7 @@ import (
 	"encoding/json"
 	"time"
 
-	"github.com/aawadallak/go-core-kit/pkg/common"
+	"github.com/aawadallak/go-core-kit/common"
 )
 
 type JobStatus string
