+++ core/job/orchestrator.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/aawadallak/go-core-kit/core/logger"
-	"github.com/aawadallak/go-core-kit/pkg/common"
+	"github.com/aawadallak/go-core-kit/common"
 )
 
 type Orchestrator struct {
