+++ core/ptr/fn.go
