+++ core/ptr/now.go
