+++ core/ptr/safe.go
