+++ core/repository/entity.go
@@ -1,6 +1,6 @@
 package repository
 
-import "github.com/aawadallak/go-core-kit/pkg/common"
+import "github.com/aawadallak/go-core-kit/common"
 
 // Entity is a type alias for common.Entity for backward compatibility.
 type Entity = common.Entity
