+++ plugin/event/publisher.go
@@ -5,7 +5,7 @@ import (
 	"context"
 
 	cevent "github.com/aawadallak/go-core-kit/core/event"
-	"github.com/aawadallak/go-core-kit/pkg/common"
+	"github.com/aawadallak/go-core-kit/common"
 	"go.opentelemetry.io/otel/trace"
 )
